core: refactor out MemPool
