+++ pkg/core/mempool/feer.go
@@ -1,4 +1,4 @@
-package core
+package mempool
 
 import (
 	"github.com/CityOfZion/neo-go/pkg/core/transaction"
