+++ pkg/core/mempool/mem_pool.go
@@ -1,4 +1,4 @@
-package core
+package mempool
 
 import (
 	"sort"
